feat(ai): add thinking/reasoning stream events and OpenAI runtime support
- Add AI_STREAM_EVENT_REASONING_START/DELTA/END to AIStreamEvent
- Update OpenAI runtime to emit reasoning events from delta content
- Update session runner to handle reasoning content in streams
- Fix agent dock UI adapter wiring

+++ editor/agent_ui/ai_agent_dock.h
@@ -91,7 +91,7 @@ class AIAgentDock : public EditorDock {
 	void _ensure_request_progress_material();
 	void _refresh_request_progress_material();
 	void _clear_request_progress_material();
-	String _format_token_count(int p_tokens) const;
+	String _format_token_count(int64_t p_tokens) const;
 
 protected:
 	static void _bind_methods();
